requirements.txt
# Telegram Bot
python-telegram-bot[ext]
python-telegram-bot[rate-limiter]

# Database ORM and adapters
SQLAlchemy
psycopg2-binary
alembic

# Yandex S3 Storage
boto3
botocore

# Cryptography
cryptography

# Environment variables
python-dotenv
# Task Scheduling
APScheduler

# Date/time parsing
python-dateutil

# Timezone support
pytz